Add enum and refactor the code
